// frontend/src/core/types/indicadores.ts
export interface IndicadoresSimulado{
    fuelCounterTA: number;
    fuelCounterTB: number;
    fuelCounterTC: number;
    fuelCounterTD: number;
    fuelCounterTotal: number;
    glpFilledNorth: number;
    glpFilledEast: number;
    glpFilledMain: number;
    glpFilledTotal: number;
    meanDeliveryTime: number; // minutos
    completedOrders: number;
    totalOrders: number;
}